Move DynamicObjectAware and DynamicObjectUtil back
Because of course we depend on them :/
