+++ subprojects/base-services-groovy/src/main/java/org/gradle/api/internal/DynamicObjectUtil.java
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.metaobject;
+package org.gradle.api.internal;
+
+import org.gradle.internal.metaobject.BeanDynamicObject;
+import org.gradle.internal.metaobject.DynamicObject;
 
 public abstract class DynamicObjectUtil {
     public static DynamicObject asDynamicObject(Object object) {
