+++ subprojects/base-services-groovy/src/main/java/org/gradle/util/ConfigureUtil.java
@@ -19,11 +19,11 @@ package org.gradle.util;
 import groovy.lang.Closure;
 import org.codehaus.groovy.runtime.GeneratedClosure;
 import org.gradle.api.Action;
+import org.gradle.api.internal.DynamicObjectUtil;
 import org.gradle.internal.Actions;
 import org.gradle.internal.metaobject.ConfigureDelegate;
 import org.gradle.internal.metaobject.DynamicInvokeResult;
 import org.gradle.internal.metaobject.DynamicObject;
-import org.gradle.internal.metaobject.DynamicObjectUtil;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
