+++ subprojects/core/src/main/java/org/gradle/api/internal/DynamicPropertyNamer.java
@@ -19,7 +19,6 @@ import groovy.lang.GroovyObject;
 import groovy.lang.MissingPropertyException;
 import org.gradle.api.Named;
 import org.gradle.api.Namer;
-import org.gradle.internal.metaobject.DynamicObjectUtil;
 
 import java.util.Map;
