+++ subprojects/core/src/main/java/org/gradle/api/internal/project/antbuilder/AntBuilderDelegate.java
@@ -21,10 +21,10 @@ import groovy.util.Node;
 import groovy.util.NodeList;
 import groovy.util.XmlParser;
 import org.apache.commons.io.IOUtils;
+import org.gradle.api.internal.DynamicObjectUtil;
 import org.gradle.internal.Cast;
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.metaobject.DynamicObject;
-import org.gradle.internal.metaobject.DynamicObjectUtil;
 
 import java.io.InputStream;
 import java.util.Collections;
